+++ src/components/GarageSearch.js
@@ -1,14 +1,28 @@
 import React, { Component } from 'react'
 import Form from 'react-bootstrap/Form'
-import Button from 'react-bootstrap/Button'
 import Container from 'react-bootstrap/Container'
 import ListadoLocalidades from './ListadoLocalidades'
 import '../styles/GarageSearch.css'
 
-
 class GarageSearch extends Component {
 
+  constructor(props){
+    super(props);
+    this.handleLocalidadChange = this.handleLocalidadChange.bind(this);
+    this.handleVehicleChange = this.handleVehicleChange.bind(this);
+  }
+  
+  handleLocalidadChange(e) {
+    this.props.onLocalidadChange(e.target.value);
+  }
+  
+  handleVehicleChange(e) {
+    this.props.onVehicleChange(e.target.value);
+  }
+
+
   render() {
+    
     return (
       <div className="searchBar">
         <Container className="form-container-search">
@@ -16,24 +30,19 @@ class GarageSearch extends Component {
             <Form.Row className="form-row-search">
               <Form.Group controlId="busquedaForm.Region" className="form-group-search">
                 <Form.Label className="d-none d-md-block label-search">Seleccione Localidad: </Form.Label>
-                <Form.Control size="sm" as="select">
+                <Form.Control size="sm" as="select" onChange={this.handleLocalidadChange}>
                   <ListadoLocalidades></ListadoLocalidades>
                 </Form.Control>
               </Form.Group>
               <Form.Group controlId="busquedaForm.VehicleType" className="form-group-search">
                 <Form.Label className="d-none d-md-block label-search">Tipo de Vehículo:</Form.Label>
-                <Form.Control size="sm" as="select">
-                  <option>Elija tipo de Vehiculo</option>
-                  <option>Autos</option>
-                  <option>Motos</option>
-                  <option>Camionetas</option>
-                  <option>Bicicletas</option>
+                <Form.Control size="sm" as="select" onChange={this.handleVehicleChange}>
+                  <option value="">Elija tipo de Vehiculo</option>
+                  <option value="autos">Autos</option>
+                  <option value="motos">Motos</option>
+                  <option value="camionetas">Camionetas</option>
+                  <option value="bicicletas">Bicicletas</option>
                 </Form.Control>
-              </Form.Group>
-              <Form.Group controlId="busquedaForm.submit" className="form-group-search">
-                <Button variant="info" type="submit" className="search-btn">
-                  Buscar
-                </Button>
               </Form.Group>
             </Form.Row>
           </Form>
@@ -43,5 +52,4 @@ class GarageSearch extends Component {
   }
 
 }
-
 export default GarageSearch