+++ src/components/ListadoLocalidades.js
@@ -23,8 +23,18 @@ class ListadoLocalidades extends React.Component {
       return (
         <>
           <option value="">Elija una localidad...</option>
-          {localidades.map((localidad) => <option value={localidad.localidad_id} 
-                                              key={localidad.localidad_id}>{localidad.nombre_localidad}</option>)}
+          {
+            localidades.map((localidad) => {
+
+              if (localidad.localidad_id !== parseInt(this.props.selected,10)) {
+                return (<option value={localidad.localidad_id}
+                  key={localidad.localidad_id}>{localidad.nombre_localidad}</option>)
+              }
+              else return (<option selected value={localidad.localidad_id}
+                key={localidad.localidad_id}>{localidad.nombre_localidad}</option>)
+            }
+            )
+          }
         </>
       )
     } else {