+++ src/components/LoadingIndicator.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import { usePromiseTracker } from "react-promise-tracker";
+import Loader from 'react-loader-spinner';
+
+const LoadingIndicator = props => {
+  const { promiseInProgress } = usePromiseTracker();
+  console.log(promiseInProgress)
+  return (
+    promiseInProgress &&
+    <div
+      style={{
+        width: "100%",
+        height: "100",
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center"
+      }}
+    >
+      <Loader type="Bars" color="#487fb9" height="100" width="100" />
+    </div>
+  );
+}
+
+export default LoadingIndicator